SFA-1   10/2/2022 1.Fixed crash issue in Stock option.(Invoice) 2.Repeat invoice completed. 3.Enter qty validation done and multiple value show in UI screen.(Primary order) 4.Remove zero in all order screen. 5.Primary screen navigate to history screen ctrl done. 6.Show total unt,crt .(primary order)

+++ app/src/main/java/com/hap/checkinproc/SFA_Activity/FPPrimaryOrderActivity.java
@@ -1447,7 +1447,7 @@ public class FPPrimaryOrderActivity extends AppCompatActivity implements View.On
                 if (CategoryType >= 0) {
 
                     holder.tvMRP.setText("₹" + ProductItem.getMRP());
-                    holder.totalQty.setText("Total Qty : " + oQty);//((Product_Details_Modalitem.get(holder.getAdapterPosition()).getQty() * (Integer.parseInt(Product_Details_Modal.getConversionFactor())))));
+                    holder.totalQty.setText("Total Qty : " + (int)oQty);//((Product_Details_Modalitem.get(holder.getAdapterPosition()).getQty() * (Integer.parseInt(Product_Details_Modal.getConversionFactor())))));
 
                     if (!ProductItem.getPImage().equalsIgnoreCase("")) {
                         holder.ImgVwProd.clearColorFilter();
